:recycle: rename package
