+++ lib/domain/settings/components/rows/menstruation.dart
